requirements.txt
asgiref==3.5.2
black==22.6.0
click==8.1.3
dj-database-url==1.0.0
Django==4.1
djangorestframework==3.13.1
geographiclib==1.52
geopy==2.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
